move utility functions/classes to util module
