+++ imgreco/end_operation.py
@@ -3,7 +3,7 @@ import sys
 import numpy as np
 from PIL import Image
 
-from richlog import get_logger
+from util.richlog import get_logger
 from . import imgops
 from . import item
 from . import minireco
