+++ imgreco/item.py
@@ -4,7 +4,7 @@ import numpy as np
 # from skimage.measure import compare_mse
 from PIL import Image
 
-from richlog import get_logger
+from util.richlog import get_logger
 from . import imgops
 from . import minireco
 from . import resources
