+++ imgreco/recruit.py
@@ -2,7 +2,7 @@ import cv2 as cv
 import numpy as np
 from PIL import Image
 
-from richlog import get_logger
+from util.richlog import get_logger
 from . import imgops
 from . import resources
 from . import util
